backend/requirements.txt
django
django-environ
djangorestframework
drf-spectacular
celery
django-summernote
django-constance[redis]
psycopg2
Pillow